requirements.txt
# Core ML and Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Deep Learning
torch==2.0.1
torchvision==0.15.2

# Data Collection
yfinance==0.2.28

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Utilities
python-dotenv==1.0.0
joblib==1.3.2

# Monitoring and Logging
prometheus-client==0.19.0
loguru==0.7.2

# Data Visualization (for notebooks)
matplotlib==3.7.2
seaborn==0.12.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Code Quality
black==23.11.0
ruff==0.1.6
mypy==1.7.0

